move selectEditorRowCol, SpriteSpeed, SpriteSize
