+++ apps/src/studio/api.js
@@ -1,20 +1,7 @@
 var constants = require('./constants');
 
-exports.SpriteSpeed = {
-  VERY_SLOW: 2,
-  SLOW: 3,
-  NORMAL: 5,
-  FAST: 8,
-  VERY_FAST: 12,
-};
-
-exports.SpriteSize = {
-  VERY_SMALL: 0.5,
-  SMALL: 0.75,
-  NORMAL: 1,
-  LARGE: 1.5,
-  VERY_LARGE: 2
-};
+exports.SpriteSpeed = constants.SpriteSpeed;
+exports.SpriteSize = constants.SpriteSize;
 
 var SPEECH_BUBBLE_TIME = 3;
