+++ apps/src/studio/constants.js
@@ -1,5 +1,21 @@
 'use strict';
 
+exports.SpriteSpeed = {
+  VERY_SLOW: 2,
+  SLOW: 3,
+  NORMAL: 5,
+  FAST: 8,
+  VERY_FAST: 12,
+};
+
+exports.SpriteSize = {
+  VERY_SMALL: 0.5,
+  SMALL: 0.75,
+  NORMAL: 1,
+  LARGE: 1.5,
+  VERY_LARGE: 2
+};
+
 exports.Direction = {
   NONE: 0,
   NORTH: 1,
